+++ config/sonic.ts
@@ -96,6 +96,7 @@ export default <NetworkData>{
                 '0x219656f33c58488d09d518badf50aa8cdcaca2aa', // wETH market 26
                 '0x5954ce6671d97d24b782920ddcdbb4b1e63ab2de', // usdc market 23
                 '0x6c49b18333a1135e9a376560c07e6d1fd0350eaf', // Ws market 28
+                '0xda14a41dbda731f03a94cb722191639dd22b35b2', // frxUSD market 37
             ],
         },
         avalon: {