+++ modules/sources/contracts/v3/vault-client.ts
@@ -3,12 +3,14 @@ import { PoolDataV3, fetchPoolData } from './fetch-pool-data';
 import { ProtocolFees, fetchProtocolFees } from './fetch-protocol-fees';
 
 export interface VaultClient {
+    viemClient: ViemClient;
     fetchPoolData: (pools: string[], blockNumber?: bigint) => Promise<{ [address: string]: PoolDataV3 }>;
     fetchProtocolFees: (blockNumber?: bigint) => Promise<ProtocolFees>;
 }
 
 export const getVaultClient = (viemClient: ViemClient, vaultAddress: string): VaultClient => {
     return {
+        viemClient,
         fetchPoolData: async (pools: string[], blockNumber?: bigint) =>
             fetchPoolData(vaultAddress, pools, viemClient, blockNumber),
         fetchProtocolFees: async (blockNumber?: bigint) => fetchProtocolFees(vaultAddress, viemClient, blockNumber),