+++ modules/sources/subgraphs/balancer-v3-vault/index.ts
@@ -14,6 +14,7 @@ import {
     PoolBalancesQueryVariables,
     PoolSnapshotFragment,
     PoolSnapshot_Filter,
+    SwapsQueryVariables,
 } from './generated/types';
 import { Chain, Prisma } from '@prisma/client';
 import { snapshotToDb } from './transformers/snapshotToDb';
@@ -24,13 +25,13 @@ export function getVaultSubgraphClient(url: string, chain: Chain) {
     return {
         ...sdk,
         chain: chain,
-        async getMetadata() {
+        async lastSyncedBlock() {
             return sdk.Metadata().then((response) => {
                 if (response && response.meta) {
-                    return response.meta;
+                    return Number(response.meta.block.number);
                 } else {
                     // Return a default value if meta is not present
-                    return Promise.reject('Error fetching metadata');
+                    return Promise.reject('Error fetching metadata block number');
                 }
             });
         },
@@ -232,6 +233,32 @@ export function getVaultSubgraphClient(url: string, chain: Chain) {
                     hasMore = false;
                 } else {
                     id = response.pools[response.pools.length - 1].id;
+                }
+            }
+
+            return data;
+        },
+        async getAllSwaps({ where, block }: SwapsQueryVariables): Promise<SwapFragment[]> {
+            const limit = 1000;
+            let hasMore = true;
+            let id = `0x`;
+            let data: SwapFragment[] = [];
+
+            while (hasMore) {
+                const response = await sdk.Swaps({
+                    where: { ...where, id_gt: id },
+                    orderBy: Swap_OrderBy.Id,
+                    orderDirection: OrderDirection.Asc,
+                    first: limit,
+                    block,
+                });
+
+                data = [...data, ...response.swaps];
+
+                if (response.swaps.length < limit) {
+                    hasMore = false;
+                } else {
+                    id = response.swaps[response.swaps.length - 1].id;
                 }
             }